refactor: extract resourcebar
